show my tour data in table also show configm alert for deleteing tour

+++ src/pages/MyList.jsx
@@ -1,9 +1,8 @@
 import { useContext, useEffect, useState } from "react";
-import { FaPeopleRobbery } from "react-icons/fa6";
-import { MdOutlineTravelExplore } from "react-icons/md";
 import { AuthContext } from "../context/AuthProvider";
 import { Link } from "react-router-dom";
-import { toast } from "react-toastify";
+
+import Swal from "sweetalert2";
 
 const MyList = () => {
   const [myTourList, setMyTourList] = useState([]);
@@ -13,24 +12,34 @@ const MyList = () => {
   const myEmail = user?.email;
 
   const handleDelete = async (id) => {
-    try {
-      const res = await fetch(`http://localhost:5000/delete-tour/${id}`, {
-        method: "DELETE",
-      });
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
 
-      const data = await res.json();
-
-      if (data.deletedCount > 0) {
-        setMyTourList((prevTour) => {
-          return prevTour.filter((tour) => {
-            return tour._id !== id;
-          });
+    if (result.isConfirmed) {
+      try {
+        const res = await fetch(`http://localhost:5000/delete-tour/${id}`, {
+          method: "DELETE",
         });
 
-        toast.success("Tour Deleted");
+        const data = await res.json();
+
+        if (data.deletedCount > 0) {
+          setMyTourList((prevTour) => {
+            return prevTour.filter((tour) => {
+              return tour._id !== id;
+            });
+          });
+        }
+      } catch (error) {
+        console.log(error);
       }
-    } catch (error) {
-      console.log(error);
     }
   };
 
@@ -54,65 +63,49 @@ const MyList = () => {
         </h2>
       )}
       <div className="flex items-center justify-center mt-8">
-        <div className="grid grid-cols-1 gap-5 md:grid-cols-2 lg:grid-cols-3">
-          {myTourList.map((tour) => {
-            return (
-              <div
-                key={tour._id}
-                className="shadow-xl card max-w-96 bg-base-100"
-              >
-                <figure>
-                  <img
-                    src={tour.photo}
-                    alt="img"
-                    className="w-full h-[220px] object-cover bg-center bg-no-repeat"
-                  />
-                </figure>
-                <div className="card-body">
-                  <div className="flex items-center justify-between w-full">
-                    <h2 className="card-title">{tour.TouristsSpotName}</h2>
-                    <h2 className="text-xl font-semibold">${tour.cost}</h2>
-                  </div>
-                  <div className="flex items-center justify-between w-full">
-                    <h4>{tour.country}</h4>
-                    <h4 className="text-sm text-gray-600">{tour.location}</h4>
-                  </div>
-                  <hr />
+        {/* table */}
 
-                  <div className="flex items-center justify-between">
-                    <h4 className="flex items-center justify-center gap-1">
-                      <FaPeopleRobbery />{" "}
-                      <span>Visitors : {tour.visitors}</span>
-                    </h4>
-                    <h4 className="flex items-center justify-center gap-1">
-                      <MdOutlineTravelExplore />{" "}
-                      <span>Travel Time : {tour.travelTime}</span>
-                    </h4>
-                  </div>
-                  <div className="flex items-center justify-between mt-2">
-                    <Link
-                      to={`/tour-details/${tour._id}`}
-                      className="px-6 text-white bg-green-500 btn hover:bg-green-600"
-                    >
-                      View
-                    </Link>
-                    <Link
-                      to={`/edit-tour/${tour._id}`}
-                      className="px-6 text-white bg-blue-500 btn hover:bg-blue-600"
-                    >
-                      Edit
-                    </Link>
-                    <button
-                      onClick={() => handleDelete(tour._id)}
-                      className="px-6 text-white bg-red-500 btn hover:bg-red-600"
-                    >
-                      Delete
-                    </button>
-                  </div>
-                </div>
-              </div>
-            );
-          })}
+        <div className="overflow-x-auto">
+          <table className="table">
+            {/* head */}
+            <thead>
+              <tr>
+                <th>Tourists Spots Name</th>
+                <th>Country</th>
+                <th>Location</th>
+                <th>Cost</th>
+                <th>Action</th>
+              </tr>
+            </thead>
+            <tbody>
+              {myTourList.map((tour) => {
+                return (
+                  <tr key={tour._id}>
+                    <td>{tour.TouristsSpotName}</td>
+                    <td>{tour.country}</td>
+                    <td>{tour.location}</td>
+                    <td>{tour.cost}</td>
+                    <td>
+                      <Link
+                        to={`/edit-tour/${tour._id}`}
+                        className="px-6 text-white bg-blue-500 btn hover:bg-blue-600"
+                      >
+                        Update
+                      </Link>
+                    </td>
+                    <td>
+                      <button
+                        onClick={() => handleDelete(tour._id)}
+                        className="px-6 text-white bg-red-500 btn hover:bg-red-600"
+                      >
+                        Delete
+                      </button>
+                    </td>
+                  </tr>
+                );
+              })}
+            </tbody>
+          </table>
         </div>
       </div>
     </section>
